+++ eCakeShop.Models/Requests/ProizvodUpdateRequest.cs
@@ -12,7 +12,7 @@ namespace eCakeShop.Models.Requests
         [Required(AllowEmptyStrings = false, ErrorMessage = "Naziv ne smije ostati prazno polje")]
         public string Naziv { get; set; }
         [Required(ErrorMessage = "Cijena ne smije ostati prazno polje")]
-        public decimal Cijena { get; set; }
+        public double Cijena { get; set; }
         [Required(ErrorMessage = "Slika je obavezna")]
         public byte[] Slika { get; set; }
         [Required]
