+++ eCakeShop.Services/Database/Proizvod.cs
@@ -15,7 +15,7 @@ namespace eCakeShop.Services.Database
         public int ProizvodID { get; set; }
         public string? Naziv { get; set; }
         public string? Sifra { get; set; }
-        public decimal Cijena { get; set; }
+        public double Cijena { get; set; }
         public byte[]? Slika { get; set; }
         public string? Opis { get; set; }
         public int VrstaProizvodaID { get; set; }
